test(core): split model tracking specs
