+++ scripts/check-file-sizes.mjs
@@ -57,7 +57,6 @@ const ALLOWLIST = new Set([
 
   // ── Rust test files ─────────────────────────────────────────────
   "crates/tracepilot-core/src/turns/tests/builders.rs",
-  "crates/tracepilot-core/src/turns/tests/model_tracking.rs",
   "crates/tracepilot-export/tests/integration.rs",
 
   // ── Pinia stores (Phase 4/5) ────────────────────────────────────
